requirements.txt
flask
flask-cors
python-dotenv
Flask-SQLAlchemy
SQLAlchemy
pymysql
opencv-python
opencv-python-headless
insightface
numpy
scipy
requests
gunicorn
onnxruntime